+++ docs/MEDS-DEV/2_datasets.mdx
@@ -2,6 +2,6 @@
 sidebar_position: 2
 ---
 
-import { Datasets } from '@site/src/components/MEDS-DEV/EntityPages';
+import { Datasets } from '@site/src/components/MEDS-DEV/entities/EntityPages';
 
 <Datasets />
