+++ docs/MEDS-DEV/3_tasks.mdx
@@ -2,6 +2,6 @@
 sidebar_position: 3
 ---
 
-import { Tasks } from '@site/src/components/MEDS-DEV/EntityPages';
+import { Tasks } from '@site/src/components/MEDS-DEV/entities/EntityPages';
 
 <Tasks />
