+++ docs/MEDS-DEV/4_models.mdx
@@ -2,6 +2,6 @@
 sidebar_position: 4
 ---
 
-import { Models } from '@site/src/components/MEDS-DEV/EntityPages';
+import { Models } from '@site/src/components/MEDS-DEV/entities/EntityPages';
 
 <Models />
