+++ docs/MEDS-DEV/5_benchmark.mdx
@@ -2,6 +2,6 @@
 sidebar_position: 5
 ---
 
-import BenchmarkPlot from '@site/src/components/MEDS-DEV/BenchmarkPlot';
+import BenchmarkPlot from '@site/src/components/MEDS-DEV/plots/BenchmarkPlot';
 
 <BenchmarkPlot />
